suspend-blocker: Make suspend durations more accurate
It seems that we should be checking for "Suspended for" messages
to provide better suspend duration stats.  If the kernel does not
emit these, tell user they need to enable CONFIG_SUSPEND_TIME=y.

Also, track the accuracy of the duration stats so we know how reliable
the data is on the histogram output.

Signed-off-by: Colin Ian King <colin.king@canonical.com>